requirements.txt
Flask==2.0.2
torch==1.9.0
torchvision==0.10.0
Pillow==8.3.1